test(project): update snapshot and test files for elements
